requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
ipython==6.2.1
-e git+git@github.com:thejessleigh/goodreads2.git#egg=goodreads2
